--- a/AUTHORS
+++ b/AUTHORS
@@ -1,0 +1,9 @@
+There several persons contribute their code to the project.
+They listed below in an alphabetical order:
+
+- Alexander (Axel) I.Grafov <grafov@gmail.com>
+- Fabrizio (Misto) Milo <mistobaan@gmail.com>
+- Jamie Stackhouse <jamie.stackhouse@redspace.com>
+- Kz26
+- Makombo
+- Zac Shenker <zshenker@brightcove.com>