--- a/AUTHORS
+++ b/AUTHORS
@@ -7,6 +7,7 @@ to the project. They listed below in an alphabetical order:
 - Bradley Falzon <brad@teambrad.net>
 - Denys Smirnov <denis.smirnov.91@gmail.com>
 - Fabrizio (Misto) Milo <mistobaan@gmail.com>
+- Hori Ryota <hori.ryota@gmail.com>
 - Jamie Stackhouse <jamie.stackhouse@redspace.com>
 - Julian Cooper <jcooper@brightcove.com>
 - Kz26