… M3U8.md
… README.md
… reader_test.go
… sample-playlists/master-with-i-frame-stream-inf.m3u8
… sample-playlists/master-with-multiple-codecs.m3u8
… sample-playlists/master-with-stream-inf-name.m3u8
… sample-playlists/media-playlist-large.m3u8
… sample-playlists/media-playlist-with-byterange.m3u8
… sample-playlists/media-playlist-with-scte35-1.m3u8
… sample-playlists/media-playlist-with-scte35.m3u8
… structure_test.go
… writer.go
… writer_test.go