@hori-ryota 2016-05-26
I want to use this. Would you merge this?