@bradleyfalzon 2016-05-26
This looks good at first glance, will need to spend some time with it, under a little time pressure with others things, maybe a week if not sooner.

It'll help me if you also describe what you did and why.

Thanks again, looking forward to reviewing it.