@hori-ryota 2016-05-29
I'm sorry, I have hurried too much.

Add more code, before
```
BenchmarkDecodeMasterPlaylist-8    30000             44103 ns/op           14958 B/op         77 allocs/op
BenchmarkDecodeMediaPlaylist-8        50          23445112 ns/op        12282139 B/op     160240 allocs/op
BenchmarkEncodeMediaPlaylist-8       100          14378131 ns/op         1749294 B/op      81643 allocs/op
```
after
```
BenchmarkDecodeMasterPlaylist-8    50000             26881 ns/op           12232 B/op         38 allocs/op
BenchmarkDecodeMediaPlaylist-8       100          16190399 ns/op         9200543 B/op      80021 allocs/op
BenchmarkEncodeMediaPlaylist-8       300           3878858 ns/op           35212 B/op        269 allocs/op
```

I'm using forked repository now, if you like this please give me an opinion.