@bradleyfalzon 2016-05-29
OK, this will take me longer to analyse. A few quick comments before 

There appears to be many improvements here, which is excellent, but some commits are just fixes to bugs caused by other commits, and also some breaking changes we almost certainly cannot accept (unexporting `MediaSegment.Duration` - and I'm not sure why).

The benchmark itself seems to require an extra 1MB, is it to run the benchmark a hundred times on a smaller example to get the same benchmark result?

It's likely we won't accept the PR without at least some rebasing, instead opting to break the changes down into more manageable chunks.

Can you describe generally the various changes you've made, for example:

- Add decoding benchmark
- Move common case statements to top for quicker decoding
- Switch to strconv.Parse* functions instead of fmt.Sscanf
- Preallocate some buffers 

Thank you, some of these changes are excellent and I look forward to merging them.