@bradleyfalzon 2016-05-30
OK,

So here's the benchmarks for each commit where the tests pass:

Before: 652e4af8ba5f7a4c7b14e0b7c560f53e6b7bbeff
```
go test -bench . -benchtime 10s -benchmem
testPASS
BenchmarkDecodeMediaPlaylist-4       500          24428571 ns/op        12279906 B/op     160060 allocs/op
BenchmarkEncodeMediaPlaylist-4      1000          13285791 ns/op         1614948 B/op      80166 allocs/op
ok      github.com/hori-ryota/m3u8      29.289s
```

Very good: f4f36c983c56a5d503b194044b4082afc3091bc2
```
go test -bench . -benchtime 10s -benchmem
testPASS
BenchmarkDecodeMediaPlaylist-4      1000          17009632 ns/op        10999885 B/op     120060 allocs/op
BenchmarkEncodeMediaPlaylist-4      1000          12839038 ns/op         1614470 B/op      80126 allocs/op
ok      github.com/hori-ryota/m3u8      32.921s
```

No substantial benefit: 52afbdf730146a76302aae3defdecbf846638836
```
go test -bench . -benchtime 10s -benchmem
testPASS
BenchmarkDecodeMediaPlaylist-4      1000          17296504 ns/op         8876106 B/op     120052 allocs/op
BenchmarkEncodeMediaPlaylist-4      1000          12718539 ns/op         1608931 B/op      80126 allocs/op
ok      github.com/hori-ryota/m3u8      33.055s
```

No substantial benefit: 78531af5c52f8949357ad466fe9a87bc51570f66
```
go test -bench . -benchtime 10s -benchmem
testPASS
BenchmarkDecodeMediaPlaylist-4      1000          17009632 ns/op        10999885 B/op     120060 allocs/op
BenchmarkEncodeMediaPlaylist-4      1000          12839038 ns/op         1614470 B/op      80126 allocs/op
ok      github.com/hori-ryota/m3u8      32.921s
```

Very good: efb283588847223639447511790331eee9c3ba56
```
go test -bench . -benchtime 10s -benchmem
testPASS
BenchmarkDecodeMediaPlaylist-4      1000          15421930 ns/op         8559427 B/op      80033 allocs/op
BenchmarkEncodeMediaPlaylist-4      1000          12535349 ns/op         1609807 B/op      80085 allocs/op
```

Additional benchmark: e89f4652aaa1d81e0eed4c809cecb94f97a14cac
```
go test -bench . -benchtime 10s -benchmem
testPASS
BenchmarkDecodeMasterPlaylist-4  1000000             23426 ns/op           14955 B/op         77 allocs/op
BenchmarkDecodeMediaPlaylist-4      1000          15394268 ns/op         8559422 B/op      80033 allocs/op
BenchmarkEncodeMediaPlaylist-4      1000          12786649 ns/op         1609807 B/op      80085 allocs/op
ok      github.com/hori-ryota/m3u8      54.993s
```

Very good `EncodeMediaPlaylist`, but regression in `DecodeMediaPlaylist` and break unexports a field? Multiple commits here and all break tests so haven't been individually benchmarked.

Results for changes between e89f4652aaa1d81e0eed4c809cecb94f97a14cac and 3629b3ce025e6cef5867017ae121e9e9dbbe5d47
```
go test -bench . -benchtime 10s -benchmem
testPASS
BenchmarkDecodeMasterPlaylist-4  1000000             22135 ns/op           14712 B/op         73 allocs/op
BenchmarkDecodeMediaPlaylist-4      1000          16163201 ns/op         9200526 B/op      80019 allocs/op
BenchmarkEncodeMediaPlaylist-4      5000           3692252 ns/op            2116 B/op         19 allocs/op
ok      github.com/hori-ryota/m3u8      59.112s
```

Very good: 5f14343349bd0f71b1745f8452fc4ea3d97fe641
```
go test -bench . -benchtime 10s -benchmem
testPASS
BenchmarkDecodeMasterPlaylist-4  1000000             10280 ns/op           12232 B/op         38 allocs/op
BenchmarkDecodeMediaPlaylist-4      1000          16184904 ns/op         9200524 B/op      80019 allocs/op
BenchmarkEncodeMediaPlaylist-4      5000           3674522 ns/op            2116 B/op         19 allocs/op
ok      github.com/hori-ryota/m3u8      47.063s
```

No substantial improvement: 8b5b7d07f07c004fdde31f18e03f71ea372208c8

```
go test -bench . -benchtime 10s -benchmem
testPASS
BenchmarkDecodeMasterPlaylist-4  1000000             10191 ns/op           12232 B/op         38 allocs/op
BenchmarkDecodeMediaPlaylist-4      1000          16525064 ns/op         9200524 B/op      80019 allocs/op
BenchmarkEncodeMediaPlaylist-4      5000           3682421 ns/op            2116 B/op         19 allocs/op
ok      github.com/hori-ryota/m3u8      47.339s
```

No substantial improvement: fe8dbb235c735869d83c3c6058c74a5c0f624178

```
go test -bench . -benchtime 10s -benchmem
testPASS
BenchmarkDecodeMasterPlaylist-4  2000000             10417 ns/op           12240 B/op         35 allocs/op
BenchmarkDecodeMediaPlaylist-4      1000          16498553 ns/op         9200524 B/op      80019 allocs/op
BenchmarkEncodeMediaPlaylist-4      5000           3666946 ns/op            2116 B/op         19 allocs/op
ok      github.com/hori-ryota/m3u8      67.854s
```