@bradleyfalzon 2016-05-30
There are some changes here that are very good, I don't want to merge them all as some performance improvements need to be weighed up against the increase risk of bugs and the readability.

I'll continue to look, but there's a lot to go through here.